apiの関数やjsonの読み書きの関数を別ファイルに記述
