+++ search/search.py
@@ -10,6 +10,8 @@ from pixivpy3 import *
 
 import tools.config as config
 import tools.pixivGetTools as pixivGetTools
+import tools.pixivApiTools as pixivApi
+import tools.jsonLoadAndWrite as jsonLoadAndWrite
 
 UTC = timezone("UTC")
 JST = timezone("Asia/Tokyo")
@@ -31,69 +33,6 @@ def getContents(link) -> json:
     contents = json.loads(contents)
 
     return contents
-
-# "いいね: likeCount, ブックマーク: bookmarkCount, 閲覧: viewCount, 投稿日時: createDate (グリニッジ標準時, +9hで日本時)"
-# class PixivGetTools:
-#     def getLikeCount(illustData, illustId) -> int:
-#         likeCount = illustData["illust"][illustId]["likeCount"]
-
-#         return int(likeCount)
-
-#     def getBookmarkCount(illustData, id) -> int:
-#         bookmarkCount = illustData["illust"][str(id)]["bookmarkCount"]
-
-#         return bookmarkCount
-
-#     def getViewCount(illustData, id) -> int:
-#         viewCount = illustData["illust"][str(id)]["viewCount"]
-
-#         return viewCount
-
-#     def getAiType(illustData, id) -> int:
-#         # not AI: 1, AI: 2
-#         aiType = illustData["illust"][str(id)]["aiType"]
-
-#         return aiType
-    
-#     def isIncludeTags(illustData, targetTags, id) -> bool:
-#         tags = illustData["illust"][str(id)]["tags"]["tags"]
-#         for tag in tags:
-#             if tag["tag"] in targetTags:
-#                 return True
-#         return False
-
-#     def isManga(illustData, id) -> bool:
-#         illustType = illustData["illust"][str(id)]["illustType"]
-#         if illustType == 1:
-#             return True
-#         return False
-
-""" 指定の日付に投稿されたイラストの端を検索 """
-def searchIllustData(api, word, sort, date) -> json:
-    # date_desc: 新しい順, date_asc: 古い順
-    json_result = api.search_illust(word=word, search_target='partial_match_for_tags', sort=sort, start_date=date, end_date=date, search_ai_type=1)
-    time.sleep(sleepTime)
-
-    return json_result
-
-# 指定日の最新のイラストIDを取得
-def getNewIllustId(api, word, date) -> int:
-    results = searchIllustData(api, word, "date_desc", date)
-    newId = results.illusts[0].id
-
-    return newId
-
-# 指定日の最古のイラストIDを取得
-def getOldIllustId(api, word, date) -> int:
-    results = searchIllustData(api, word, "date_asc", date)
-    oldId = results.illusts[0].id
-
-    return oldId
-
-def getIllustData(api, illustId) -> json:
-    illustData = api.illust_detail(illustId)
-    time.sleep(sleepTime)
-    return illustData
 
 def searchDownload(api, id, detailData):
     dict = {}
@@ -140,21 +79,6 @@ def searchDownload(api, id, detailData):
 
     print("id {} is downloaded".format(id))
 
-""" jsonファイルからidを読み込み、ダウンロード"""
-def downloadFromJson(api, detailData):
-    for id in detailData:
-        searchDownload(api, id, detailData)
-
-def openJson(jsonDir) -> json:
-    with open(jsonDir, "r") as f:
-        data = json.load(f)
-
-    return data
-
-def saveJson(jsonDir, data):
-    with open(jsonDir, "w") as f:
-        json.dump(data, f)
-
 """ ログイン用 """
 def apiLogin() -> AppPixivAPI:
     api = AppPixivAPI()
@@ -168,12 +92,12 @@ def main():
     yesterday = datetime.now() - timedelta(days=1)
     date_str = yesterday.strftime('%Y-%m-%d')
 
-    detailData = openJson(illustJsonDir)
-    searched = openJson(searchedJsonDir)
+    detailData = jsonLoadAndWrite.openJson(illustJsonDir)
+    searched = jsonLoadAndWrite.openJson(searchedJsonDir)
 
     dataMaxId = searched["id"] + 1
     
-    minId = max(getOldIllustId(api, "", date_str), dataMaxId)
+    minId = max(pixivApi.getOldIllustId(api, "", date_str), dataMaxId)
 
     # for id in range(minId, minId + maxCount):
     #     searchDownload(api, id, detailData)
